requirements.txt
# 核心依赖
PySide6>=6.5.0
pyserial>=3.5
numpy>=1.24.0

# 数据处理与分析
pandas>=2.0.0

# 可选依赖 - 光谱仪功能
nidaqmx>=0.6.5
pyqtgraph>=0.13.0
scipy>=1.10.0

# 开发依赖（可选）
pytest>=7.4.0
pytest-qt>=4.2.0
mypy>=1.5.0
black>=23.7.0
flake8>=6.1.0

